+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 ## 0.1.0
 Initial version with Dimensionless, Mass and Volume quantities.
+
+## 0.1.1
+Lenght quantities
+Misspelling fixes.
+Better mass usage example.